pyproject.toml
# MathForge packaging and tool configuration (PEP 621).
# Install: pip install -e ".[dev]"   (editable + dev dependencies)

[build-system]
# Build backend for creating wheels / editable installs.
requires = ["setuptools>=75", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mathforge-agent"
version = "0.2.0"
description = "LangGraph + Claude math agent with sandboxed Python execution"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    # Agent runtime: LangGraph prebuilt ReAct + Anthropic chat model.
    "langgraph==1.1.8",
    "langchain-anthropic==1.4.1",
    "langchain-core==1.3.0",
    "python-dotenv==1.2.2",
    # Scientific stack: imported in sandbox preamble for model-generated code.
    "numpy>=2.0,<3",
    "sympy>=1.13,<2",
    "matplotlib>=3.9,<4",
    "scipy>=1.14,<2",
]

[project.optional-dependencies]
# Lint + test; not needed for `pip install .` in production.
dev = [
    "pytest>=8.3",
    "pytest-asyncio>=0.25",
    "ruff>=0.8",
]

[project.scripts]
# After install: `mathforge` on PATH calls `main.main()`.
mathforge = "main:main"

[tool.setuptools]
# Flat layout: top-level modules, not a `src/` package.
py-modules = ["agent", "config", "main", "sandbox", "tools"]

[tool.pytest.ini_options]
# pytest-asyncio: async tests without explicit loop fixtures.
asyncio_mode = "auto"
testpaths = ["tests"]
filterwarnings = [
    # LangGraph deprecations / upstream noise during test runs.
    "ignore:create_react_agent has been moved:DeprecationWarning",
    "ignore:'asyncio.iscoroutinefunction' is deprecated:DeprecationWarning",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
# E/F: pycodestyle/pyflakes; I: isort; UP: pyupgrade.
select = ["E", "F", "I", "UP"]
